python/requirements.txt
torch
stable-baselines3
gymnasium
pandas
polars
supabase
python-dotenv
requests
